Moved context to to CartContext
