--- react-context/my-app/src/components/CartContent.tsx
@@ -1,13 +1,0 @@
-import { createContext } from 'react';
-
-export type CartValue = {
-  cart: Product[];
-  addToCart: (product: Product) => void;
-};
-
-const defaultCartValue: CartValue = {
-  cart: [],
-  addToCart: () => undefined,
-};
-
-export const CartContext = createContext(defaultCartValue);
