+++ react-context/my-app/src/components/useCart.ts
@@ -1,0 +1,9 @@
+import { CartContext } from './CartContext';
+import { useContext } from 'react';
+
+export function useCart() {
+  const context = useContext(CartContext);
+  if (!context)
+    throw new Error('useCart must be used inside of the CartProvider');
+  return context;
+}
