+++ react-context/my-app/src/pages/ProductDetails.tsx
@@ -1,7 +1,7 @@
-import { useEffect, useState, useContext } from 'react';
+import { useEffect, useState } from 'react';
 import { Link, useNavigate, useParams } from 'react-router-dom';
 import { readProduct, type Product, toDollars } from '../lib';
-import { CartContext } from '../components/CartContent';
+import { useCart } from '../components/useCart.ts';
 
 export function ProductDetails() {
   const { productId } = useParams();
@@ -9,7 +9,7 @@ export function ProductDetails() {
   const [product, setProduct] = useState<Product>();
   const [isLoading, setIsLoading] = useState(true);
   const [error, setError] = useState<unknown>();
-  const { addToCart } = useContext(CartContext);
+  const { addToCart } = useCart();
 
   useEffect(() => {
     async function loadProduct(productId: number) {
